added simd function if avx/neon/v are not present

+++ bench/label/sub.cpp
@@ -8,7 +8,7 @@ B63_BASELINE(sub, nn) {
 	kAryLabel v1{}, v2{}, v3{};
 	uint64_t k_lower, k_higher;
 	B63_SUSPEND {
-		for (int i = 0; i < n; ++i) {
+		for (uint64_t i = 0; i < n; ++i) {
 			v1.data()[i] = i;
 			v2.data()[i] = i;
 		}
@@ -16,7 +16,7 @@ B63_BASELINE(sub, nn) {
 		translate_level(&k_lower, &k_higher, -1, __level_translation_array);
 	}
 
-	int32_t res = 0;
+	uint64_t res = 0;
 
 	for (; res < nn; res++) {
 		kAryLabel::sub(v3, v1, v2, k_lower, k_higher);
